Changed cuda code time measurement. Increased occupancy for rendering
But fps now 2 times lower, but cuda occupancy is 55% instead of 11%
And in time measurements I removed memcpy and cudaMalloc from function and now it approx 100 times faster

+++ include/config.h
@@ -1,7 +1,7 @@
 #ifndef CONFIG_H__
 #define CONFIG_H__
 
-typedef float mandel_t;
+typedef double mandel_t;
 
 const int WIDTH = 800;
 const int HEIGHT = 600;
